Moved stuff to UI and removed "type" attribute from Experiment
