+++ app/src/main/res/layout/fragment_my_exp.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MyExpFragment" >
+    tools:context=".ui.MyExpFragment" >
 
     <ListView
         android:id="@+id/my_exp_list_view"
